+++ boggle/breaker.py
@@ -1,14 +1,23 @@
+"""Break a board class (prove it has no boards with >P points) using EvalTree.
+
+This is the 2025 strategy, as described at:
+https://www.danvk.org/2025/02/13/boggle2025.html
+
+This performs several "lift" operations on the tree to synchronize choices across
+subtrees and reduce the bound. After a few of these (the number is the "switchover_level"),
+it stops modifying the tree and does a DFS on the remaining choices until the board class
+is fully broken (this done in "bound_remaining_boards").
+
+Lifting uses a lot of memory but can make bound_remaining_boards dramatically faster,
+especially for complex, hard-to-break board classes. On the other hand, lifting too many
+times wastes memory and is slower than bound_remaining_boards. There is a sweet spot for
+each board class.
+"""
+
 import dataclasses
 import time
 from collections import Counter, defaultdict
 from dataclasses import dataclass
-from typing import Sequence
-
-from cpp_boggle import (
-    BucketBoggler33,
-    BucketBoggler34,
-    BucketBoggler44,
-)
 
 from boggle.boggler import PyBoggler
 from boggle.eval_tree import (
@@ -17,8 +26,6 @@ from boggle.eval_tree import (
 )
 from boggle.letter_grouping import get_letter_map, reverse_letter_map, ungroup_letters
 from boggle.split_order import SPLIT_ORDER
-
-type BucketBoggler = BucketBoggler33 | BucketBoggler34 | BucketBoggler44
 
 
 @dataclass
@@ -37,19 +44,10 @@ class BreakDetails:
 
 
 @dataclass
-class IBucketBreakDetails(BreakDetails):
-    by_level: Counter[int]
-
-    def asdict(self):
-        d = super().asdict()
-        d["by_level"] = dict(self.by_level.items())
-        return d
-
-
-@dataclass
 class HybridBreakDetails(BreakDetails):
     sum_union: int
     bounds: dict[int, int]
+    nodes: dict[int, int]
     boards_to_test: int
     expanded_to_test: int
     init_nodes: int
@@ -57,129 +55,6 @@ class HybridBreakDetails(BreakDetails):
     freed_nodes: int
     free_time_s: float
     num_filtered: dict[int, int]
-
-
-class IBucketBreaker:
-    """Break by recursively splitting cells and evaluating with ibuckets.
-
-    This is simple and allocates no memory, but does lots of repeated work.
-    """
-
-    def __init__(
-        self,
-        boggler: BucketBoggler,
-        dims: tuple[int, int],
-        best_score: int,
-        *,
-        num_splits: int,
-    ):
-        self.bb = boggler
-        self.best_score = best_score
-        self.details_ = None
-        self.elim_ = 0
-        self.orig_reps_ = 0
-        self.dims = dims
-        self.split_order = SPLIT_ORDER[dims]
-        self.num_splits = num_splits
-
-    def SetBoard(self, board: str):
-        return self.bb.ParseBoard(board)
-
-    def Break(self) -> IBucketBreakDetails:
-        self.details_ = IBucketBreakDetails(
-            num_reps=0,
-            elapsed_s=0.0,
-            failures=[],
-            by_level=Counter(),
-            elim_level=Counter(),
-            secs_by_level=defaultdict(float),
-            # root_score_bailout=None,
-        )
-        self.elim_ = 0
-        self.orig_reps_ = self.bb.NumReps()
-        start_time_s = time.time()
-        self.AttackBoard()
-
-        self.details_.elapsed_s = time.time() - start_time_s
-        self.details_.num_reps = self.orig_reps_
-        # TODO: debug output
-        return self.details_
-
-    def PickABucket(self, level: int):
-        pick = -1
-        splits = []
-
-        cells = self.bb.as_string().split(" ")
-
-        # TODO: lots of comments in C++ source with ideas for exploration here.
-        for order in self.split_order:
-            if len(cells[order]) > 1:
-                pick = order
-                break
-
-        if pick == -1:
-            return pick, splits, cells
-
-        cell = cells[pick]
-        cell_len = len(cell)
-        if cell_len == 26:
-            # TODO: is this ever useful?
-            splits = ["bdfgjqvwxz", "aeiou", "lnrsy", "chkmpt"]
-            # splits = ["aeiou", "sy", "bdfgjkmpvwxzq", "chlnrt"]
-        elif cell_len >= 9:
-            splits = ["".join(split) for split in even_split(cell, self.num_splits)]
-
-        else:
-            splits = [*cell]
-
-        return pick, splits, cells
-
-    def SplitBucket(self, level: int) -> None:
-        cell, splits, cells = self.PickABucket(level)
-        if cell == -1:
-            # it's just a board
-            board = "".join(cells)
-            # print(f"Unable to break board: {board}")
-            self.details_.failures.append(board)
-            return
-
-        for i, split in enumerate(splits):
-            bd = [*cells]
-            bd[cell] = split
-            # C++ version uses ParseBoard() + Cell() here.
-            assert self.bb.ParseBoard(" ".join(bd))
-            self.AttackBoard(level + 1, i + 1, len(splits))
-
-    def AttackBoard(self, level: int = 0, num: int = 1, out_of: int = 1) -> None:
-        # TODO: debug output
-        # reps = self.bb.NumReps()
-
-        self.details_.by_level[level] += 1
-        start_s = time.time()
-        ub = self.bb.UpperBound(self.best_score)
-        elapsed_s = time.time() - start_s
-        self.details_.secs_by_level[level] += elapsed_s
-        if level == 0:
-            # self.details_.root_score_bailout = (ub, self.bb.Details().bailout_cell)
-            self.details_.root_bound = ub
-        if ub <= self.best_score:
-            self.elim_ += self.bb.NumReps()
-            # self.details_.max_depth = max(self.details_.max_depth, level)
-            self.details_.elim_level[level] += 1
-        else:
-            self.SplitBucket(level)
-
-
-# TODO: there's probably a more concise way to express this.
-# Or maybe not https://stackoverflow.com/a/54802737/388951
-def even_split[T](xs: Sequence[T], num_buckets: int) -> list[list[T]]:
-    splits = [[]]
-    length = len(xs)
-    for i, x in enumerate(xs):
-        if num_buckets * i >= (len(splits) * length):
-            splits.append([])
-        splits[-1].append(x)
-    return splits
 
 
 class HybridTreeBreaker:
@@ -197,7 +72,7 @@ class HybridTreeBreaker:
         *,
         # TODO: ideally this should depend on the node_count of the tree.
         switchover_level: int,
-        free_after_score: bool,
+        free_after_lift: bool,
         log_breaker_progress: bool,
         letter_grouping: str = "",
     ):
@@ -210,7 +85,7 @@ class HybridTreeBreaker:
         self.dims = dims
         self.split_order = SPLIT_ORDER[dims]
         self.switchover_level = switchover_level
-        self.free_after_score = free_after_score
+        self.free_after_lift = free_after_lift
         self.log_breaker_progress = log_breaker_progress
         self.rev_letter_grouping = (
             reverse_letter_map(get_letter_map(letter_grouping))
@@ -241,6 +116,7 @@ class HybridTreeBreaker:
             boards_to_test=0,
             expanded_to_test=0,
             init_nodes=0,
+            nodes={},
             total_nodes=0,
             num_filtered={},
             freed_nodes=0,
@@ -262,6 +138,7 @@ class HybridTreeBreaker:
         self.details_.bounds[0] = tree.bound
         self.details_.sum_union = self.etb.Details().sum_union
         self.details_.init_nodes = arena.num_nodes()
+        self.details_.nodes[0] = self.details_.init_nodes
 
         self.AttackTree(tree, 1, arena)
         self.details_.elapsed_s = time.time() - start_time_s
@@ -297,9 +174,9 @@ class HybridTreeBreaker:
                 self.details_.num_filtered[level] = n_filtered
         if self.log_breaker_progress:
             self.mark += 1
-            print(
-                f"{level=} {cell=} {tree.bound=}, {tree.unique_node_count(self.mark)} unique nodes"
-            )
+            count = tree.unique_node_count(self.mark)
+            print(f"{level=} {cell=} {tree.bound=}, {count} unique nodes")
+            self.details_.nodes[level] = count
 
         self.AttackTree(tree, level + 1, arena)
 
@@ -316,7 +193,7 @@ class HybridTreeBreaker:
     def switch_to_score(self, tree: EvalNode, level: int, arena) -> None:
         # This reduces the amount of time we use max memory, but it's a ~5% perf hit.
         # start_s = time.time()
-        if not self.free_after_score:
+        if self.free_after_lift:
             self.mark += 1
             start_s = time.time()
             self.details_.freed_nodes = arena.mark_and_sweep(tree, self.mark)
